--- exporter/clickhousetracesexporter/README.md
@@ -1,7 +1,0 @@
-Archive storage?
-hold/cold to clickhouse? Use of S3 for cold?
-
-support other configs/flags?
-Is Operations Table needed?
-
-cat ./opentelemetry-collector/exporter/clickhouseexporter/sql-schema/signoz-index.sql |  sudo ./clickhouse client -h 18.220.17.59 -mn
