+++ exporter/clickhousetracesexporter/factory.go
@@ -16,8 +16,12 @@ package clickhousetracesexporter
 
 import (
 	"context"
+	"fmt"
 	"time"
 
+	"github.com/ClickHouse/clickhouse-go/v2"
+	"github.com/google/uuid"
+	"github.com/jellydator/ttlcache/v3"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config/configretry"
 	"go.opentelemetry.io/collector/exporter"
@@ -26,9 +30,7 @@ import (
 
 const (
 	// The value of "type" key in configuration.
-	typeStr          = "clickhousetraces"
-	primaryNamespace = "clickhouse"
-	archiveNamespace = "clickhouse-archive"
+	typeStr = "clickhousetraces"
 )
 
 func createDefaultConfig() component.Config {
@@ -59,7 +61,47 @@ func createTracesExporter(
 ) (exporter.Traces, error) {
 
 	c := cfg.(*Config)
-	oce, err := newExporter(cfg, params.Logger, params)
+
+	client, err := newClickhouseClient(ctx, c)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create clickhouse client: %w", err)
+	}
+
+	id := uuid.New()
+
+	// keys cache is used to avoid duplicate inserts for the same attribute key.
+	keysCache := ttlcache.New[string, struct{}](
+		ttlcache.WithTTL[string, struct{}](240*time.Minute),
+		ttlcache.WithCapacity[string, struct{}](50000),
+	)
+	go keysCache.Start()
+
+	// resource fingerprint cache is used to avoid duplicate inserts for the same resource fingerprint.
+	// the ttl is set to the same as the bucket rounded value i.e 1800 seconds.
+	// if a resource fingerprint is seen in the bucket already, skip inserting it again.
+	rfCache := ttlcache.New[string, struct{}](
+		ttlcache.WithTTL[string, struct{}](distributedTracesResourceV2Seconds*time.Second),
+		ttlcache.WithDisableTouchOnHit[string, struct{}](),
+		ttlcache.WithCapacity[string, struct{}](100000),
+	)
+	go rfCache.Start()
+
+	meter := params.MeterProvider.Meter("github.com/SigNoz/signoz-otel-collector/exporter/clickhousetracesexporter")
+	writerOpts := []WriterOption{
+		WithClickHouseClient(client),
+		WithLogger(params.Logger),
+		WithMeter(meter),
+		WithKeysCache(keysCache),
+		WithRFCache(rfCache),
+		WithAttributesLimits(c.AttributesLimits),
+		WithExporterID(id),
+	}
+
+	exporterOpts := []TraceExporterOption{
+		WithNewUsageCollector(id, client),
+	}
+
+	exporter, err := newExporter(c, params, writerOpts, exporterOpts)
 	if err != nil {
 		return nil, err
 	}
@@ -68,9 +110,30 @@ func createTracesExporter(
 		ctx,
 		params,
 		cfg,
-		oce.pushTraceData,
-		exporterhelper.WithShutdown(oce.Shutdown),
+		exporter.pushTraceData,
+		exporterhelper.WithShutdown(exporter.Shutdown),
 		exporterhelper.WithTimeout(c.TimeoutConfig),
 		exporterhelper.WithQueue(c.QueueConfig),
 		exporterhelper.WithRetry(c.BackOffConfig))
+}
+
+func newClickhouseClient(ctx context.Context, cfg *Config) (clickhouse.Conn, error) {
+	options, err := clickhouse.ParseDSN(cfg.Datasource)
+	if err != nil {
+		return nil, err
+	}
+	// setting maxIdleConnections = numConsumers + 1 to avoid `prepareBatch:clickhouse: acquire conn timeout` error
+	maxIdleConnections := cfg.QueueConfig.NumConsumers + 1
+	if options.MaxIdleConns < maxIdleConnections {
+		options.MaxIdleConns = maxIdleConnections
+		options.MaxOpenConns = maxIdleConnections + 5
+	}
+	db, err := clickhouse.Open(options)
+	if err != nil {
+		return nil, err
+	}
+	if err := db.Ping(ctx); err != nil {
+		return nil, err
+	}
+	return db, nil
 }
